vacuum implemented
